Rework whole ui structure, delegation work and model adjustments. 1. Event Adapter 2. Partial View Model 3. Document and DB logging
