+++ src/main/java/Main.java
@@ -1,18 +1,22 @@
-import controller.EventManager;
-import controller.MainController;
-import controller.ViewManager;
+import client.controls.EventManager;
+import client.controls.MainController;
+import client.controls.ViewManager;
 import javafx.application.Application;
 import javafx.stage.Stage;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
-import persistence.CalendarDAO;
+import server.persistence.CalendarDAO;
 
-/**
- * Name: Mr.Chatchapol Rasameluangon
- * ID:   5810404901
+import java.time.format.DateTimeFormatter;
+
+/*
+ * @author Chatchapol Rasameluangon
+ * id: 5810404901
  */
 
 public class Main extends Application {
+    private static DateTimeFormatter defaultDatePattern = DateTimeFormatter.ofPattern("dd-MM-yyyy");
+    private static String defaultColor = "#814567";
     private CalendarDAO dao;
     private ApplicationContext context;
