--- src/main/java/common/search/Searcher.java
@@ -1,10 +1,0 @@
-package common.search;
-
-import javafx.collections.ObservableList;
-import model.Event;
-
-public interface Searcher {
-
-    ObservableList<Event> search(ObservableList<Event> list, String text);
-
-}
