+++ src/main/resources/css/MFXTheme.css
@@ -66,7 +66,7 @@
     -fx-background-radius: 0, 0, 0, 0;
     -fx-padding: 0em 0.166667em 0em 0.166667em; /* 0 2px 0 2px*/
 
-    -fx-font-family: "Segoe UI", Helvetica, Arial, sans-serif;
+    -fx-font-family: Arial, Helvetica, sans-serif;
     -fx-font-size: 1.25em; /* 16 */
 }
 
@@ -474,6 +474,8 @@
     -fx-background-insets: -1, 2;
     -fx-background-radius: 0, 0;
 }
+
+
