+++ src/main/resources/fxml/MonthView.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.text.*?>
 <BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0"
             prefWidth="700.0" style="-fx-background-color: #202020;" stylesheets="@../css/MainTheme.css"
-            xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="view.MonthView">
+            xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="client.ui.monthview.MonthView">
     <center>
         <GridPane fx:id="gridView" BorderPane.alignment="CENTER">
             <columnConstraints>
