+++ src/main/resources/main-context.xml
@@ -4,8 +4,8 @@
        xmlns:context="http://www.springframework.org/schema/context"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
 
-    <context:component-scan base-package="persistence"/>
-    <context:component-scan base-package="controller"/>
+    <context:component-scan base-package="server.persistence"/>
+    <context:component-scan base-package="client.controls"/>
 
     <bean id = "dataSource" class="org.sqlite.SQLiteDataSource">
         <property name= "url" value="jdbc:sqlite:#{ systemProperties['user.home'] }/CalendarDB/Events.db"/>
